requirements.txt
Flask
requests
beautifulsoup4
nltk
transformers
torch
pandas
pymongo
kafka-python
